refactor: move navigation input to more logical folder
